pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skymarket"
version = "0.1.0"
description = "Minimal safe live-trading Polymarket weather bot"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "py-clob-client>=0.34.6",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

